docs: add documentation on how to create an activity page
chore: code cleanup

test: undo changes to test
