+++ bciers/apps/reporting/src/app/bceidbusiness/industry_user/activities/general_stationary_combustion/page.tsx
@@ -5,7 +5,7 @@ import { actionHandler } from "@bciers/actions";
 export default async function Page() {
   const reportDate = "2024-04-01"; // This should be passed in once we have a path to this page from starting a report
   const activityData = await actionHandler(
-    `reporting/get-initial-activity-data?activity_name=General stationary combustion&report_date=${reportDate}`,
+    `reporting/initial-activity-data?activity_name=General stationary combustion&report_date=${reportDate}`,
     "GET",
     "",
   );
